dcm framework 64-bit bug
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@5423 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCM Framework/DCMLimitedObject.h
@@ -26,8 +26,8 @@
 - (id)initWithData:(NSData *)data lastGroup:(unsigned short)lastGroup;
 - (id)initWithContentsOfFile:(NSString *)file lastGroup:(unsigned short)lastGroup;
 - (id)initWithContentsOfURL:(NSURL *)aURL lastGroup:(unsigned short)lastGroup;
-- (id)initWithDataContainer:(DCMDataContainer *)data lengthToRead:(long)lengthToRead byteOffset:(long  *)byteOffset characterSet:(DCMCharacterSet *)characterSet lastGroup:(unsigned short)lastGroup;
+- (id)initWithDataContainer:(DCMDataContainer *)data lengthToRead:(int)lengthToRead byteOffset:(int*)byteOffset characterSet:(DCMCharacterSet *)characterSet lastGroup:(unsigned short)lastGroup;
 
-- (long)readDataSet:(DCMDataContainer *)dicomData toGroup:(unsigned short)lastGroup byteOffset:(long *)byteOffset;
+- (int)readDataSet:(DCMDataContainer *)dicomData toGroup:(unsigned short)lastGroup byteOffset:(int *)byteOffset;
 
 @end
