+++ DCM Framework/DCMObject.h
@@ -152,7 +152,7 @@
 * @param characterSet The DCMCharacterSet used for decoding
 * @param decodePixelData Flag to decode contained pixelData
 */
-- (id)initWithDataContainer:(DCMDataContainer *)data lengthToRead:(long)lengthToRead byteOffset:(long  *)byteOffset characterSet:(DCMCharacterSet *)characterSet decodingPixelData:(BOOL)decodePixelData;
+- (id)initWithDataContainer:(DCMDataContainer *)data lengthToRead:(int)lengthToRead byteOffset:(int  *)byteOffset characterSet:(DCMCharacterSet *)characterSet decodingPixelData:(BOOL)decodePixelData;
 
 /** Empty initializer */
 - (id)init;
@@ -182,17 +182,17 @@
 
 /** Parse the dataset\n
 * Used when parsing the DICOM data. */
-- (long)readDataSet:(DCMDataContainer *)dicomData lengthToRead:(long)lengthToRead byteOffset:(long *)byteOffset;
+- (int)readDataSet:(DCMDataContainer *)dicomData lengthToRead:(int)lengthToRead byteOffset:(int *)byteOffset;
 
 /** Parse of Sequence attribute\n
 * Used when parsing the DICOM data. */
-- (long)readNewSequenceAttribute:(DCMAttribute *)attr dicomData:(DCMDataContainer *)dicomData byteOffset:(long *)byteOffset lengthToRead:(long)lengthToRead specificCharacterSet:(DCMCharacterSet *)specificCharacterSet;
+- (int)readNewSequenceAttribute:(DCMAttribute *)attr dicomData:(DCMDataContainer *)dicomData byteOffset:(int *)byteOffset lengthToRead:(int)lengthToRead specificCharacterSet:(DCMCharacterSet *)specificCharacterSet;
 
 /** Create a DCMAttribute\n
 * Used when parsing the DICOM data. */
 - (DCMAttribute *) newAttributeForAttributeTag:(DCMAttributeTag *)tag 
 			vr:(NSString *)vr 
-			length:(long) vl 
+			length:(int) vl 
 			data:(DCMDataContainer *)dicomData 
 			specificCharacterSet:(DCMCharacterSet *)specificCharacterSet
 			isExplicit:(BOOL) explicitTS
